feat: 투두 유효성 검사 함수 생성, 리드미 파일
