+++ src/api/auth.js
@@ -9,13 +9,3 @@ export const signUp = async body => {
   const authSignUp = await instance.post('/auth/signup', body);
   return authSignUp;
 };
-
-export const emailCheck = email => {
-  let regExp = /[@]/;
-  return email.trim() !== '' && email !== 'undefined' && regExp.test(email);
-};
-
-export const passwordCheck = password => {
-  let regExp = /.{8,}/;
-  return password.trim() !== '' && password !== 'undefined' && regExp.test(password);
-};
