+++ src/utils/todoValidation.js
@@ -1,0 +1,3 @@
+export const validateTodo = todo => {
+  return todo.replace(/ /g, '').length >= 3;
+};
